+++ android/app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
             android:exported="true" />
 
         <service
-            android:name=".mediacontrol.PlaybackService"
+            android:name=".mediasession.PlaybackService"
             android:exported="true"
             android:foregroundServiceType="mediaPlayback">
             <intent-filter>
